Move log-summary-parser.ts to /log-insights
