+++ extensions/ql-vscode/src/log-insights/log-summary-parser.ts
