+++ extensions/ql-vscode/src/query-evaluation-logging/eval-log-tree-builder.ts
@@ -1,5 +1,5 @@
 import { ChildEvalLogTreeItem, EvalLogTreeItem } from "./eval-log-viewer";
-import { EvalLogData as EvalLogData } from "../pure/log-summary-parser";
+import { EvalLogData as EvalLogData } from "../log-insights/log-summary-parser";
 
 /** Builds the tree data for the evaluator log viewer for a single query run. */
 export class EvalLogTreeBuilder {
